refactoring, classes from com.softgroup.authorization.api.message splitted into two subpackages
