+++ cources.server/softgroup/services/authorization/authorization-api/src/main/java/com/softgroup/authorization/api/message/request/SmsConfirmRequestData.java
@@ -1,4 +1,4 @@
-package com.softgroup.authorization.api.message;
+package com.softgroup.authorization.api.message.request;
 
 import com.softgroup.common.protocol.RequestData;
